Add required fixes
